chore: Types for ResponsiveChecklist
