+++ apps/editor.planx.uk/src/@planx/components/Option/model.tsx
@@ -7,7 +7,7 @@ import { Rule } from "../shared/RuleBuilder/types";
  * The Option node doesn't have a direct Editor or Public interface
  *  - Editor: Controlled and authored via the parent Question component
  *  - Public: Viewed and selected via the parent Question component
- * 
+ *
  * XXX: Maps to ComponentType.Answer
  */
 export interface Option {
@@ -30,6 +30,9 @@ export interface ConditionalOption extends Option {
     rule: Rule;
   };
 }
+
+export type AnyOption = Option | ConditionalOption;
+export type AnyOptions = Option[] | ConditionalOption[];
 
 export const optionValidationSchema = object({
   id: string(),
