+++ apps/editor.planx.uk/src/@planx/components/shared/BaseChecklist/Editor/components/ExpandableSwitch.tsx
@@ -1,0 +1,36 @@
+import { ComponentType } from "@opensystemslab/planx-core/types";
+import { toggleExpandableChecklist } from "@planx/components/shared/BaseChecklist/model";
+import React from "react";
+import { Switch } from "ui/shared/Switch";
+
+import { Props } from "../types";
+
+export const TypeNarrowedExpandableSwitch: React.FC<Props> = ({
+  formik,
+  disabled,
+  type,
+}) => {
+  if (type === ComponentType.Checklist)
+    return (
+      <Switch
+        checked={!!formik.values.groupedOptions}
+        onChange={() =>
+          formik.setValues(toggleExpandableChecklist(formik.values))
+        }
+        label="Expandable"
+        disabled={disabled}
+      />
+    );
+
+  if (type === ComponentType.ResponsiveChecklist)
+    return (
+      <Switch
+        checked={!!formik.values.groupedOptions}
+        onChange={() =>
+          formik.setValues(toggleExpandableChecklist(formik.values))
+        }
+        label="Expandable"
+        disabled={disabled}
+      />
+    );
+};
